må runde av grad før utbetalingslinjer bygges
fordi spenn har konvertert grad til int uten avrundning  må eksisterende utbetalingslinjer rundes ned,
fordi vi må reflektere den verdien som OS har fått.

+++ sykepenger-model/src/main/kotlin/no/nav/helse/person/DelegatedPersonVisitor.kt
@@ -986,7 +986,7 @@ internal class DelegatedPersonVisitor(private val delegateeFun: () -> PersonVisi
         satstype: Satstype,
         beløp: Int?,
         aktuellDagsinntekt: Int?,
-        grad: Double?,
+        grad: Int?,
         delytelseId: Int,
         refDelytelseId: Int?,
         refFagsystemId: String?,
