+++ sykepenger-model/src/main/kotlin/no/nav/helse/serde/api/builders/OppdragBuilder.kt
@@ -88,7 +88,7 @@ internal class OppdragBuilder : BuilderState() {
         satstype: Satstype,
         beløp: Int?,
         aktuellDagsinntekt: Int?,
-        grad: Double?,
+        grad: Int?,
         delytelseId: Int,
         refDelytelseId: Int?,
         refFagsystemId: String?,
