+++ sykepenger-model/src/main/kotlin/no/nav/helse/utbetalingstidslinje/Feriepengeberegner.kt
@@ -268,7 +268,7 @@ internal class Feriepengeberegner(
                 satstype: Satstype,
                 beløp: Int?,
                 aktuellDagsinntekt: Int?,
-                grad: Double?,
+                grad: Int?,
                 delytelseId: Int,
                 refDelytelseId: Int?,
                 refFagsystemId: String?,
