+++ sykepenger-model/src/test/kotlin/no/nav/helse/spleis/e2e/RevurderInntektTest.kt
@@ -850,7 +850,7 @@ private fun Utbetalingslinje.assertUtbetalingslinje(
             satstype: Satstype,
             beløp: Int?,
             aktuellDagsinntekt: Int?,
-            grad: Double?,
+            grad: Int?,
             delytelseId: Int,
             refDelytelseId: Int?,
             refFagsystemId: String?,
